requirements.txt
robotframework==6.0.1
robotframework-pabot==2.15.0
robotframework-seleniumlibrary==6.0.0
openpyxl==3.0.10
pydrive~=1.3.1
google-auth-oauthlib
selenium==4.7.2
requests~=2.32.0
GitPython~=3.1.43
